Apply code review changes
