+++ editor/src/messages/input_mapper/utility_types/misc.rs
@@ -4,6 +4,7 @@ use crate::messages::input_mapper::utility_types::input_keyboard::{KeyStates, NU
 use crate::messages::input_mapper::utility_types::input_mouse::NUMBER_OF_MOUSE_BUTTONS;
 use crate::messages::prelude::*;
 
+use core::time::Duration;
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone)]
@@ -144,5 +145,24 @@ impl ActionKeys {
 				String::new()
 			}
 		}
+	}
+}
+
+#[derive(Copy, Clone, Debug, Default, Eq, PartialEq, Serialize, Deserialize)]
+pub struct FrameTimeInfo {
+	timestamp: Duration,
+	prev_timestamp: Option<Duration>,
+}
+
+impl FrameTimeInfo {
+	pub fn frame_duration(&self) -> Option<Duration> {
+		self.prev_timestamp.map(|prev| self.timestamp - prev)
+	}
+
+	pub fn advance_timestamp(&mut self, next_timestamp: Duration) {
+		debug_assert!(next_timestamp >= self.timestamp);
+
+		self.prev_timestamp = Some(self.timestamp);
+		self.timestamp = next_timestamp;
 	}
 }
