+++ editor/src/messages/input_mapper/utility_types/mod.rs
@@ -1,5 +1,4 @@
 pub mod input_keyboard;
 pub mod input_mouse;
-pub mod input_time;
 pub mod macros;
 pub mod misc;
